+++ Module.php
@@ -47,7 +47,7 @@ class Module extends \yii\base\Module implements BootstrapInterface
      */
     public $logTarget;
     /**
-     * @var array list of debug panels. The array keys are the panel IDs, and values are the corresponding
+     * @var array|Panel[] list of debug panels. The array keys are the panel IDs, and values are the corresponding
      * panel class names or configuration arrays. This will be merged with [[corePanels()]].
      * You may reconfigure a core panel via this property by using the same panel ID.
      * You may also disable a core panel by setting it to be false in this property.